UI-13 update(accordion): rename context file
